+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+- TimeZone:
+  - Migrated xTimeZone resource from [xTimeZone](https://github.com/PowerShell/xTimeZone)
+    and renamed to TimeZone - fixes [Issue #157](https://github.com/PowerShell/ComputerManagementDsc/issues/157).
+- Moved Test-Command from ComputerManagementDsc.ResourceHelper to
+  ComputerManagementDsc.Common module to match what TimeZone requires.
+  It was not exported in ComputerManagementDsc.ResourceHelper and not
+  used.
+
 ## 5.0.0.0
 
 - BREAKING CHANGE: